+++ tests/dependencies/test_locales.py
@@ -30,14 +30,26 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+
+from unittest import mock
+
 import hikari
 
-from tanjun.dependencies import locales
+import tanjun
 
 
 class TestBasicLocaliser:
+    def test_add_to_client(self):
+        client = tanjun.Client(mock.AsyncMock())
+        localiser = tanjun.dependencies.BasicLocaliser()
+
+        assert localiser.add_to_client(client) is None
+
+        assert client.injector.get_type_dependency(tanjun.dependencies.AbstractLocaliser) is localiser
+        assert client.injector.get_type_dependency(tanjun.dependencies.AbstractLocalizer) is localiser
+
     def test_get_all_variants(self):
-        localiser = locales.BasicLocaliser().set_variants(
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
             "yeet",
             {
                 hikari.Locale.BG: "BIGGIE CHEESE",
@@ -57,7 +69,7 @@ class TestBasicLocaliser:
         }
 
     def test_get_all_variants_ignores_default(self):
-        localiser = locales.BasicLocaliser().set_variants(
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
             "ayanami",
             {hikari.Locale.ES_ES: "que?", hikari.Locale.DA: "Danken", "default": "yeet", hikari.Locale.FR: "for real"},
         )
@@ -69,7 +81,7 @@ class TestBasicLocaliser:
         }
 
     def test_get_all_variants_when_format(self):
-        localiser = locales.BasicLocaliser().set_variants(
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
             "rei",
             {
                 hikari.Locale.DA: "ok {no} meow",
@@ -87,15 +99,17 @@ class TestBasicLocaliser:
         }
 
     def test_get_all_variants_when_not_known(self):
-        assert locales.BasicLocaliser().get_all_variants("nope") == {}
+        assert tanjun.dependencies.BasicLocaliser().get_all_variants("nope") == {}
 
     def test_localise(self):
-        localiser = locales.BasicLocaliser().set_variants("the end", {hikari.Locale.FI: "bye"}, vi="echo", fi="meow")
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
+            "the end", {hikari.Locale.FI: "bye"}, vi="echo", fi="meow"
+        )
 
         assert localiser.localise("the end", "vi") == "echo"
 
     def test_localise_when_format(self):
-        localiser = locales.BasicLocaliser().set_variants(
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
             "ANDER",
             {hikari.Locale.EN_GB: "nyaa", hikari.Locale.FR: "ashley {now}", hikari.Locale.IT: "op"},
             moomin="okokokok",
@@ -104,20 +118,22 @@ class TestBasicLocaliser:
         assert localiser.localise("ANDER", hikari.Locale.FR, now="444") == "ashley 444"
 
     def test_localise_when_tag_unknown(self):
-        localiser = locales.BasicLocaliser().set_variants("NERV", {hikari.Locale.BG: "nom"})
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants("NERV", {hikari.Locale.BG: "nom"})
 
         assert localiser.localise("NERV", hikari.Locale.EN_GB) is None
 
     def test_localise_when_id_unknown(self):
-        assert locales.BasicLocaliser().localise("foo", hikari.Locale.FR) is None
+        assert tanjun.dependencies.BasicLocaliser().localise("foo", hikari.Locale.FR) is None
 
     def test_localize(self):
-        localiser = locales.BasicLocaliser().set_variants("the end", {hikari.Locale.FI: "bye"}, vi="echo", fi="meow")
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
+            "the end", {hikari.Locale.FI: "bye"}, vi="echo", fi="meow"
+        )
 
         assert localiser.localize("the end", "vi") == "echo"
 
     def test_localize_when_format(self):
-        localiser = locales.BasicLocaliser().set_variants(
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants(
             "ANDER",
             {hikari.Locale.EN_GB: "nyaa", hikari.Locale.FR: "ashley {now}", hikari.Locale.IT: "op"},
             moomin="okokokok",
@@ -126,9 +142,9 @@ class TestBasicLocaliser:
         assert localiser.localize("ANDER", hikari.Locale.FR, now="444") == "ashley 444"
 
     def test_localize_when_tag_unknown(self):
-        localiser = locales.BasicLocaliser().set_variants("NERV", {hikari.Locale.BG: "nom"})
+        localiser = tanjun.dependencies.BasicLocaliser().set_variants("NERV", {hikari.Locale.BG: "nom"})
 
         assert localiser.localize("NERV", hikari.Locale.EN_GB) is None
 
     def test_localize_when_id_unknown(self):
-        assert locales.BasicLocaliser().localize("foo", hikari.Locale.FR) is None
+        assert tanjun.dependencies.BasicLocaliser().localize("foo", hikari.Locale.FR) is None